backend/requirements.txt
# FastAPI core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.0.0

# AI/ML
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.21.0

# OCR for text extraction
pytesseract>=0.3.10
opencv-python>=4.8.0

# CORS and utilities
starlette>=0.27.0
